unification: resolve merge conflicts

+++ gaios/Transaction/TransactionViewController.swift
@@ -36,7 +36,7 @@ class TransactionViewController: UIViewController {
             return WalletManager.current?.currentSession?.isResetActive ?? false ||
             !transaction.canRBF || account?.isWatchonly ?? false
     }
-    private var amounts: [(key: String, value: UInt64)] {
+    private var amounts: [(key: String, value: Int64)] {
         if transaction?.type == .some(.redeposit) {
             return []
         }
@@ -44,7 +44,7 @@ class TransactionViewController: UIViewController {
         // OUT transactions in BTC/L-BTC have fee included
         if transaction.type == .some(.outgoing) {
             let feeAsset = session?.gdkNetwork.getFeeAsset()
-            amounts = amounts.map { $0.key == feeAsset ? ($0.key, $0.value - transaction.fee) : $0 }
+            amounts = amounts.map { $0.key == feeAsset ? ($0.key, $0.value - Int64(transaction.fee)) : $0 }
         }
         return amounts.filter({ $0.value != 0 })
     }
